apps/transferservice/requirements.txt
alembic==1.4.3;
amqp==2.6.1;
babel==2.9.0;
beautifulsoup4==4.9.3
billiard==3.6.3.0;
blinker==1.4
cachetools==4.1.1;
celery==4.4.7;
certifi==2020.11.8;
chardet==3.0.4;
click==7.1.2;
cloudscraper==1.2.48
feedparser==6.0.2;
flask-babel==2.0.0
flask-caching==1.9.0
flask-jwt-extended==3.25.0
flask-mail==0.9.1
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4;
flask-testing==0.8.0
flask-wtf==0.14.3
flask==1.1.2;
flower==0.9.5
# grpcio==1.34.0;
humanize==3.1.0;
idna==2.10;
importlib-metadata==3.1.1;
itsdangerous==1.1.0;
jinja2==2.11.2;
kombu==4.6.11;
libcst==0.3.15;
mako==1.1.3;
markupsafe==1.1.1;
marshmallow==3.9.1;
mypy-extensions==0.4.3;
numpy==1.19.4;
pandas==1.1.4;
prometheus-client==0.8.0
proto-plus==1.11.0;
protobuf==3.14.0;
pyasn1-modules==0.2.8;
pyasn1==0.4.8;
pyjwt==1.7.1
pymysql==0.10.1
pyparsing==2.4.7;
python-dateutil==2.8.1;
python-dotenv==0.15.0
python-editor==1.0.4;
pytz==2020.4;
pyyaml==5.3.1;
redis
requests-toolbelt==0.9.1
requests==2.25.0;
rsa==4.6;
sgmllib3k==1.0.0;
six==1.15.0;
soupsieve==2.0.1;
sqlalchemy==1.3.20;
tornado==6.1;
typing-extensions==3.7.4.3;
typing-inspect==0.6.0;
urllib3==1.26.2;
uwsgi==2.0.19.1
vine==1.3.0;
werkzeug==1.0.1;
wtforms==2.3.3
zipp==3.4.0;
psycopg2;
pybase62;
flask-script;